Add a palette of predefined patterns
Use the 1-3 keys to select a predefined pattern to place where you
click. Use the R key to rotate the pattern 90 degrees counterclockwise.

A preview of the pattern is drawn in yellow at the mouse position.
Crosshairs are also drawn to help locate the mouse when the pattern is
small.