Try to improve performance
The image was rendered by calling SDL_RenderFillRect for every live
cell. This was assumed to be slow. Also, because only live cells were
rendered, the performance was variable depending on how many cells were
live.

In an attempt to improve performance, the 1280x720 image is now drawn as
a single texture in software and blitted to the window with a single
SDL_RenderCopy call. Also, all cells are rendered so that I can optimise
for the worse case. (Dead cells are rendered in the background color.)

However, performance is not measurably better. It may be worse.

+RTS -s shows a large increase in the rate of allocations (930 MB per
MUT second up from 680 MB). A run with the profiler revealed that
checkNeighbour and stepGeneration account for a large portion of the MUT
time and allocations. The next optimisation efforts should be focussed
on improving the performance of stepGeneration.