Try to improve performance more
A couple of things were changed to try to improve performance:

1. Upgrade to GHC 8.8.

2. Use primitive arrays.

The mutable unboxed vectors for the generation state have been replaced
with arrays from the "primitive" package. This API is largely unsafe,
notably the "readCell" and "writeCell" functions no longer perform
bounds checking and the caller is now responsible. This resulted in a
small speed boost due to fewer allocations, but according to the
profile, "readCell" and "countNeighbours" are is still a source of many
allocations. I'm not sure how to proceed to reduce allocations further.

3. More precise sleep.

The Rust version is able to maintain a much more stable timing so the
sleep logic has been changed to match that version more closely. It
helped a little, but the Haskell version still fluctuates a few
milliseconds per frame.

Overall, the gains are decent but not amazing. The program is still
unable to push much further beyond 60 Hz on my machine. Allocations that
I argue are unnecessary are still occurring in the generation step.

+RTS -s reports that the allocation rate has increased to 2 GB per MUT
second, so there seems to be some potential gains to be found in forcing
more strictness...